Use a StatelessSession (through Jakarta Data) for DB access
Because our scenario is simple CRUD, without any association,
it does not leverage the persistence context at all:
we always do just one DB operation for every HTTP request.

For that reason, **in this scenario**,
a `StatelessSession` has a good chance to perform better:
it'll do the same operations in DB, but will not try to maintain
a persistence context and will thus avoid all associated overhead
(mainly memory allocations).

Note we kind of bypass Panache at the moment because StatelessSession
will only be available in [Panache 2](https://github.com/quarkusio/quarkus/issues/46096).
So for now we use Jakarta Data, which is a perfectly acceptable
alternative in this specific scenario.